Refactor buffer editor code
