+++ mfp/gui/imgui/buffer_editor/__init__.py
@@ -1,0 +1,1 @@
+from .buffer_editor import BufferEditor
