+++ mfp/gui/imgui/buffer_editor/buffer_editor.py
@@ -1,15 +1,9 @@
 """
 buffer_editor.py -- BufferEditor() class definition
 """
-import asyncio
-import os
 from datetime import datetime
-from posix_ipc import SharedMemory
-import numpy as np
 from imgui_bundle import implot, imgui
 from mfp import log
-from mfp.buffer_info import BufferInfo
-from .app_window import menu_bar
 
 
 def fmt_time(ttime):
@@ -75,391 +69,13 @@ class BufferEditor:
     def set_playhead_at_pointer(self):
         self.implot_playhead_needs_set = True
 
-    async def init_working_patch(self):
-        from mfp.gui_main import MFPGUI
-        if self.working_patch_id:
-            await self.close_working_patch()
-
-        self.working_patch_id = await MFPGUI().mfp.open_file(None, show_gui=False)
-        self.working_patch_info = await MFPGUI().mfp.get_tooltip_info(self.working_patch_id, details=True)
-        buffer_params = dict(
-            channels=self.buffer_info.channels + 2,
-            size=self.buffer_info.size,
-        )
-
-        self.working_source_info = await MFPGUI().mfp.create(
-            "buffer~",
-            ", ".join([f"{key}={value!r}" for key, value in buffer_params.items()]),
-            self.working_patch_info.get("name"),
-            None,
-            "source_buffer"
-        )
-        self.working_source_id = self.working_source_info.get("obj_id")
-
-        # wait for buffer to be initialized
-        source = None
-        while source is None:
-            try:
-                all_buffers = await MFPGUI().mfp.get_buffer_info()
-                source = next(b for b in all_buffers if b.get("proc_name") == "source_buffer")
-            except StopIteration:
-                await asyncio.sleep(0.1)
-
-        source_buf = source.get("buf_info")
-        self.working_buf_id = source_buf.buf_id
-        self.working_buf_obj = SharedMemory(source_buf.buf_id)
-        buffer_params["buf_id"] = source_buf.buf_id
-        buffer_params["channels"] = source_buf.channels
-        buffer_params["size"] = source_buf.size
-
-        self.working_sink_info = await MFPGUI().mfp.create(
-            "buffer~",
-            ", ".join([f"{key}={value!r}" for key, value in buffer_params.items()]),
-            self.working_patch_info.get("name"),
-            None,
-            "sink_buffer"
-        )
-        self.working_sink_id = self.working_sink_info.get("obj_id")
-
-        self.working_aud0_info = await MFPGUI().mfp.create(
-            "out~", "0", self.working_patch_info.get("name"), None, "audition 0"
-        )
-        self.working_aud1_info = await MFPGUI().mfp.create(
-            "out~", "1", self.working_patch_info.get("name"), None, "audition 1"
-        )
-        await MFPGUI().mfp.connect(self.working_sink_id, 0, self.working_aud0_info.get("obj_id"), 0)
-        await MFPGUI().mfp.connect(self.working_sink_id, 1, self.working_aud1_info.get("obj_id"), 0)
-
-        self.buffer_sync(self.shm_obj, self.buffer_info, self.working_buf_obj, source_buf)
-        self.buffer_compute_peaks()
-
-    async def close_working_patch(self):
-        from mfp.gui_main import MFPGUI
-
-        if self.working_patch_id:
-            await MFPGUI().mfp.delete(self.working_patch_id)
-            self.working_patch_id = None
-
-    ########################################
-    # fx patch
-    async def fx_open_patch(self, fxname):
-        from mfp.gui_main import MFPGUI
-
-        # create new patch wrapping the specified effect
-        self.fx_patch_id = await MFPGUI().mfp.open_file(
-            "bufedit.fx~.mfp",
-            initargs=f"'{fxname}', {self.buffer_info.channels}",
-            show_gui=True,
-        )
-
-        # each fx patch needs to take a number param for channel count
-        # input level and xfade channels on the left
-
-        # connect source to fx
-        for port in range(self.buffer_info.channels):
-            await MFPGUI().mfp.connect(
-                self.working_source_id, port, self.fx_patch_id, port + 2
-            )
-            await MFPGUI().mfp.connect(
-                self.fx_patch_id, port, self.working_sink_id, port
-            )
-
-            if port % 2 == 0:
-                # disconnect previous outputs
-                await MFPGUI().mfp.disconnect(
-                    self.working_sink_id, port, self.working_aud0_info.get("obj_id"), 0
-                )
-                await MFPGUI().mfp.connect(
-                    self.fx_patch_id, port, self.working_aud0_info.get("obj_id"), 0
-                )
-            else:
-                await MFPGUI().mfp.disconnect(
-                    self.working_sink_id, port, self.working_aud1_info.get("obj_id"), 0
-                )
-                await MFPGUI().mfp.connect(
-                    self.fx_patch_id, port, self.working_aud1_info.get("obj_id"), 0
-                )
-
-        # connect input level and xfade
-        for port in [0, 1]:
-            await MFPGUI().mfp.connect(
-                self.working_source_id, self.buffer_info.channels + port,
-                self.fx_patch_id, port
-            )
-
-        for element in [
-            'apply_button', 'cancel_button', 'selection_toggle',
-            'bypass_toggle', 'xfade_enum', 'preroll_enum'
-        ]:
-            element_id = await MFPGUI().mfp.resolve(element, self.fx_patch_id)
-            self.fx_patch_elements[element] = MFPGUI().objects.get(element_id)
-
-        # add actions for Apply and Cancel buttons
-        cancel = self.fx_patch_elements["cancel_button"]
-        cancel.extra_action = self.fx_close_patch
-
-        apply = self.fx_patch_elements["apply_button"]
-        apply.extra_action = self.fx_apply_patch
-
-    async def fx_apply_patch(self):
-        from mfp.gui_main import MFPGUI
-        rec_channels = sum(
-            1 << c for c in range(self.buffer_info.channels)
-        )
-
-        source_params = dict(
-            buf_mode=5,
-            play_channels=0xff,
-            buf_pos=0,
-            region_start=0,
-            region_end=self.buffer_info.size,
-            trig_trigger=1,
-        )
-        sink_params = dict(
-            buf_mode=0,
-            play_channels=0,
-            rec_channels=rec_channels,
-            rec_enabled=1,
-            buf_pos=0,
-            region_start=0,
-            region_end=self.buffer_info.size,
-            trig_trigger=1,
-        )
-
-        def fw_handler(target, signal, status):
-            if status == 0:
-                log.debug(f"[bufedit] re-grabbing data from {self.working_buf_obj}")
-                self.buffer_grab(self.working_buf_obj)
-
-        MFPGUI().appwin.signal_listen("freewheel", fw_handler)
-        await MFPGUI().mfp.send(self.working_source_id, 0, source_params)
-        await MFPGUI().mfp.send(self.working_sink_id, 0, sink_params)
-        await MFPGUI().mfp.freewheel(self.working_sink_id, self.buffer_info.size)
-
-    async def fx_close_patch(self):
-        from mfp.gui_main import MFPGUI
-
-        # remove actions for Apply and Cancel buttons
-        cancel = self.fx_patch_elements["cancel_button"]
-        cancel.extra_action = None
-
-        # reconnect sink buffer to audition outs, disconnect FX outs
-        for port in range(self.buffer_info.channels):
-            await MFPGUI().mfp.disconnect(
-                self.working_source_id, port,
-                self.fx_patch_id, port + 2
-            )
-            await MFPGUI().mfp.disconnect(
-                self.fx_patch_id, port,
-                self.working_sink_id, port
-            )
-
-            if port % 2 == 0:
-                # reconnect previous outputs
-                await MFPGUI().mfp.connect(
-                    self.working_sink_id, port, self.working_aud0_info.get("obj_id"), 0
-                )
-                await MFPGUI().mfp.disconnect(
-                    self.fx_patch_id, port, self.working_aud0_info.get("obj_id"), 0
-                )
-            else:
-                await MFPGUI().mfp.connect(
-                    self.working_sink_id, port, self.working_aud1_info.get("obj_id"), 0
-                )
-                await MFPGUI().mfp.disconnect(
-                    self.fx_patch_id, port, self.working_aud1_info.get("obj_id"), 0
-                )
-
-        # connect input level and xfade
-        for port in [0, 1]:
-            await MFPGUI().mfp.disconnect(
-                self.working_source_id, self.buffer_info.channels + port,
-                self.fx_patch_id, port
-            )
-
-        # delete FX patch
-        await MFPGUI().mfp.delete(self.fx_patch_id)
-        self.fx_patch_id = None
-
-    ########################################
-    # buffer operations
-    def buffer_grab(self, shm_obj=None):
-        def offset(channel):
-            return channel * self.buffer_info.size * self.FLOAT_SIZE
-
-        if self.buffer_info is None:
-            return None
-
-        if shm_obj is None:
-            if self.shm_obj is None:
-                self.shm_obj = SharedMemory(self.buffer_info.buf_id)
-            shm_obj = self.shm_obj
-
-        self.buffer_data = []
-        self.channel_selections = [None] * (self.buffer_info.channels + 1)
-        self.channel_selections_active = [False] * (self.buffer_info.channels + 1)
-        self.implot_limits = None
-        self.implot_limits_need_set = [None] * (self.buffer_info.channels + 1)
-
-        try:
-            for c in range(self.buffer_info.channels):
-                os.lseek(shm_obj.fd, offset(c), os.SEEK_SET)
-                slc = os.read(shm_obj.fd, int(self.buffer_info.size * self.FLOAT_SIZE))
-                self.buffer_data.append(np.fromstring(slc, dtype=np.float32))
-        except Exception as e:
-            log.debug("[grab]: error grabbing data", e)
-            import traceback
-            traceback.print_exc()
-            return None
-        self.buffer_compute_peaks()
-
-    def buffer_sync(self, from_obj, from_info, to_obj, to_info, data=None):
-        sync_channels = to_info.channels
-        if from_info:
-            sync_channels = min(from_info.channels, to_info.channels)
-        for c in range(sync_channels):
-            self.buffer_sync_channel(c, from_obj, from_info, to_obj, to_info, data)
-
-    def buffer_sync_channel(self, channel, from_obj, from_info, to_obj, to_info, data=None):
-        def offset(buf_info, channel):
-            return channel * buf_info.size * self.FLOAT_SIZE
-
-        if from_obj:
-            os.lseek(from_obj.fd, offset(from_info, channel), os.SEEK_SET)
-            slc = os.read(from_obj.fd, int(from_info.size * self.FLOAT_SIZE))
-        elif data is not None:
-            slc = data.tobytes()
-        else:
-            slc = self.buffer_data[channel].tobytes()
-
-        os.lseek(to_obj.fd, offset(to_info, channel), os.SEEK_SET)
-        os.write(to_obj.fd, slc)
-
-    def buffer_compute_peaks(self):
-        self.buffer_peaks = {}
-        padding = 10 - len(self.buffer_data[0]) % 10
-        padded = [
-            np.pad(chan, (0, padding), mode='constant')
-            for chan in self.buffer_data
-        ]
-        total_time = len(padded[0]) / self.buffer_info.rate
-        sample_time = 1/self.buffer_info.rate
-        self.implot_total_time = total_time
-        self.implot_limits = implot.Rect(
-            x_min=0, x_max=total_time, y_min=-1, y_max=1
-        )
-        self.implot_limits_need_set = [True] * (self.buffer_info.channels + 1)
-        self.buffer_peaks["1"] = (
-            padded,
-            np.arange(0, total_time, sample_time, dtype=np.float32)
-        )
-        last_peaks = padded
-
-        for peak_factor in (10, 100, 1000, 10000):
-            if peak_factor == 10:
-                shape = 10
-            else:
-                shape = 20
-            next_peaks = []
-            for channel in last_peaks:
-                maxima = channel.reshape(-1, shape).max(axis=1)
-                minima = channel.reshape(-1, shape).min(axis=1)
-                combined = np.ravel(np.column_stack((maxima, minima)))
-                padding = 20 - len(combined) % 20
-                padded = np.pad(combined, (0, padding), mode='constant')
-                next_peaks.append(padded)
-
-            x_values = np.arange(
-                0, sample_time * peak_factor * len(padded) / 2,
-                peak_factor * sample_time / 2, dtype=np.float32
-            )
-            self.buffer_peaks[str(peak_factor)] = (
-                next_peaks, x_values
-            )
-            last_peaks = next_peaks
-
-    def get_peak_scale(self):
-        """
-        called within a begin_plot()
-        """
-        limits = implot.get_plot_limits()
-        compress = self.buffer_info.rate * (
-            (max(limits.x.max, 1.0) - limits.x.min)
-            / self.app_window.canvas_panel_width
-        )
-
-        if compress < 10:
-            peak_scale = "1"
-        elif compress < 100:
-            peak_scale = "10"
-        elif compress < 1000:
-            peak_scale = "100"
-        elif compress < 10000:
-            peak_scale = "1000"
-        else:
-            peak_scale = "10000"
-        return peak_scale
-
-    def buffer_set_selection(self):
-        preroll = 0
-        xfade = 0
-
-        if "preroll_enum" in self.fx_patch_elements:
-            preroll = self.fx_patch_elements.get('preroll_enum').value
-            xfade = self.fx_patch_elements.get('xfade_enum').value
-
-        working_buf_info = BufferInfo(
-            buf_id=self.working_buf_id,
-            size=self.buffer_info.size,
-            channels=self.buffer_info.channels + 2,
-            rate=self.buffer_info.rate,
-            offset=self.buffer_info.offset
-        )
-        # input level signal
-        startpos = int(
-            (self.implot_selection.x.min - (preroll / 1000.0)) * self.buffer_info.rate
-        )
-        endpos = int(
-            (self.implot_selection.x.max + (preroll / 1000.0)) * self.buffer_info.rate
-        )
-
-        startpos = max(0, startpos)
-        endpos = min(self.buffer_info.size, endpos)
-
-        input_arr = np.zeros(self.buffer_info.size, dtype=np.float32)
-        input_arr[startpos:endpos] = 1
-        self.buffer_sync_channel(
-            self.buffer_info.channels, None, None, self.working_buf_obj, working_buf_info,
-            data=input_arr
-        )
-
-        # crossfade signal
-        inramp_start = int(self.implot_selection.x.min * self.buffer_info.rate)
-        outramp_end = int(self.implot_selection.x.max * self.buffer_info.rate)
-
-        ramp_len = min(
-            int((xfade / 1000) * self.buffer_info.rate),
-            int((outramp_end - inramp_start) / 2)
-        )
-
-        inramp_end = inramp_start + ramp_len
-        outramp_start = outramp_end - ramp_len
-
-        xfade_arr = np.zeros(self.buffer_info.size, dtype=np.float32)
-        xfade_arr[inramp_start:outramp_end] = 1
-        xfade_arr[inramp_start:inramp_end] = np.linspace(0, 1, ramp_len, dtype=np.float32)
-        xfade_arr[outramp_start:outramp_end] = np.linspace(1, 0, ramp_len, dtype=np.float32)
-
-        self.buffer_sync_channel(
-            self.buffer_info.channels + 1, None, None, self.working_buf_obj, working_buf_info,
-            data=xfade_arr
-        )
-
     ########################################
     # toolbar
     def render_toolbar(self):
         from mfp.gui import image_utils
         from mfp.gui_main import MFPGUI
+        from . import menu_button
+
         line_height = imgui.get_text_line_height()
         button_size = 1.25*line_height
 
@@ -675,7 +291,7 @@ class BufferEditor:
             imgui.open_popup("##bufedit_popup")
 
         imgui.pop_style_var(2)
-        menu_bar.render_bufedit_menu(self.app_window)
+        menu_button.render_bufedit_menu(self.app_window)
         imgui.end()
         return 2 * button_size
 
@@ -1023,3 +639,7 @@ class BufferEditor:
 
         self.implot_playhead_start_time = datetime.now()
         self.implot_playhead_start_pos = self.implot_playhead
+
+from . import buffer_ops
+from . import working_patch
+from . import fx_patch
