+++ setup.py
@@ -11,7 +11,7 @@ setup (name = 'mfp',
        packages = [
            'mfp', 'mfp.builtins',
            'mfp.gui', 'mfp.gui.modes',
-           'mfp.gui.imgui', 'mfp.gui.imgui.app_window',
+           'mfp.gui.imgui', 'mfp.gui.imgui.app_window', 'mfp.gui.imgui.buffer_editor',
            'mfp.gui.clutter', 'mfp.gui.clutter.xyplot'
        ],
        entry_points = { 'console_scripts': ['mfpmain=mfp.mfp_main:main_sync_wrapper',
